refactor(rendering): Tier 4 data-oriented optimization
Split monolithic rendering classes into focused components:

Task 4.3 - Split LightDrawer into three classes:
- LightFBO: owns framebuffer + texture with grow-only resize
- LightShader: owns shader program, VAO/VBO, and SSBO
- LightDrawer: thin orchestrator coordinating both

Task 4.2 - Split TooltipDrawer into three classes:
- TooltipCollector: header-only pool for per-frame tooltip data (no NanoVG dependency)
- TooltipRenderer: NanoVG rendering of tooltip cards
- NVGImageCache: shared sprite image cache with context change detection
- TooltipData: extracted pure data types (no rendering dependency)

Task 4.1 - Type-safe resident image tracking:
- Replace vector<void*> with vector<Image*> in SpriteDatabase
- Remove static_cast<Image*> from TextureGarbageCollector
- Remove friend class TooltipDrawer from GameSprite
