+++ source/rendering/ui/tooltip_data.h
@@ -15,22 +15,13 @@
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 //////////////////////////////////////////////////////////////////////
 
-#ifndef RME_TOOLTIP_DRAWER_H_
-#define RME_TOOLTIP_DRAWER_H_
+#ifndef RME_TOOLTIP_DATA_H_
+#define RME_TOOLTIP_DATA_H_
 
-#include "app/definitions.h"
-#include <iostream>
-#include "map/position.h"
-#include "rendering/core/render_view.h"
-#include <vector>
-#include <string>
+#include <cstdint>
 #include <string_view>
-#include <sstream>
-#include <unordered_map>
-
-class Item;
-class Waypoint;
-struct NVGcontext;
+#include <vector>
+#include "map/position.h"
 
 struct ContainerItem {
 	uint16_t id;
@@ -102,11 +93,9 @@ struct TooltipData {
 	}
 
 	void clear() {
-		// Reset scalars
 		pos = Position();
 		category = TooltipCategory::ITEM;
 		itemId = 0;
-		// clear strings
 		itemName = {};
 		actionId = 0;
 		uniqueId = 0;
@@ -118,72 +107,6 @@ struct TooltipData {
 		containerItems.clear();
 		containerCapacity = 0;
 	}
-};
-
-class TooltipDrawer {
-public:
-	TooltipDrawer();
-	~TooltipDrawer();
-
-	// Add a structured tooltip for an item
-	void addItemTooltip(const TooltipData& data);
-	void addItemTooltip(TooltipData&& data);
-
-	// Request a tooltip object from the pool. Call commitTooltip() to finalize.
-	TooltipData& requestTooltipData();
-	void commitTooltip();
-
-	// Add a waypoint tooltip
-	void addWaypointTooltip(Position pos, std::string_view name);
-
-	// Draw all tooltips
-	void draw(NVGcontext* vg, const ViewState& view);
-
-	// Clear all tooltips
-	void clear();
-
-protected:
-	struct FieldLine {
-		std::string_view label;
-		std::string_view value;
-		uint8_t r, g, b;
-		std::vector<std::string_view> wrappedLines; // For multi-line values
-	};
-	std::vector<FieldLine> scratch_fields;
-	size_t scratch_fields_count = 0;
-	std::string storage; // Scratch buffer for text generation
-
-	std::vector<TooltipData> tooltips;
-	size_t active_count = 0;
-
-	std::unordered_map<uint32_t, int> spriteCache; // sprite_id -> nvg image handle
-	NVGcontext* lastContext = nullptr;
-
-	// Helper to get or load sprite image
-	int getSpriteImage(NVGcontext* vg, uint16_t itemId);
-
-	// Helper to get header color based on category
-	void getHeaderColor(TooltipCategory cat, uint8_t& r, uint8_t& g, uint8_t& b) const;
-
-	// Refactored drawing helpers
-	struct LayoutMetrics {
-		float width;
-		float height;
-		float valueStartX;
-		float gridSlotSize;
-		int containerCols;
-		int containerRows;
-		float containerHeight;
-		int totalContainerSlots;
-		int emptyContainerSlots;
-		int numContainerItems;
-	};
-
-	void prepareFields(const TooltipData& tooltip);
-	LayoutMetrics calculateLayout(NVGcontext* vg, const TooltipData& tooltip, float maxWidth, float minWidth, float padding, float fontSize);
-	void drawBackground(NVGcontext* vg, float x, float y, float width, float height, float cornerRadius, const TooltipData& tooltip);
-	void drawFields(NVGcontext* vg, float x, float y, float valueStartX, float lineHeight, float padding, float fontSize);
-	void drawContainerGrid(NVGcontext* vg, float x, float y, const TooltipData& tooltip, const LayoutMetrics& layout);
 };
 
 #endif
