+++ source/rendering/utilities/light_drawer.cpp
@@ -17,64 +17,24 @@
 
 #include "app/main.h"
 #include "rendering/utilities/light_drawer.h"
+#include "rendering/utilities/light_fbo.h"
+#include "rendering/utilities/light_shader.h"
 #include "rendering/utilities/light_calculator.h"
 #include <glm/gtc/matrix_transform.hpp>
-#include <format>
-#include <limits>
-#include "map/tile.h"
-#include "game/item.h"
 #include "rendering/core/drawing_options.h"
 #include "rendering/core/render_view.h"
 #include "rendering/core/gl_scoped_state.h"
 
-// GPULight struct moved to header
-
-LightDrawer::LightDrawer() {
-}
-
-LightDrawer::~LightDrawer() {
-	// Resources are RAII managed
-}
-
-void LightDrawer::InitFBO() {
-	fbo = std::make_unique<GLFramebuffer>();
-	fbo_texture = std::make_unique<GLTextureResource>(GL_TEXTURE_2D);
-
-	// Initial dummy size
-	ResizeFBO(32, 32);
-
-	glNamedFramebufferTexture(fbo->GetID(), GL_COLOR_ATTACHMENT0, fbo_texture->GetID(), 0);
-
-	GLenum status = glCheckNamedFramebufferStatus(fbo->GetID(), GL_FRAMEBUFFER);
-	if (status != GL_FRAMEBUFFER_COMPLETE) {
-		spdlog::error("LightDrawer FBO Incomplete: {}", status);
-	}
-}
-
-void LightDrawer::ResizeFBO(int width, int height) {
-	if (width == buffer_width && height == buffer_height) {
-		return;
-	}
-
-	buffer_width = width;
-	buffer_height = height;
-
-	glTextureStorage2D(fbo_texture->GetID(), 1, GL_RGBA8, width, height);
-
-	// Set texture parameters
-	glTextureParameteri(fbo_texture->GetID(), GL_TEXTURE_MIN_FILTER, GL_NEAREST);
-	glTextureParameteri(fbo_texture->GetID(), GL_TEXTURE_MAG_FILTER, GL_NEAREST);
-	glTextureParameteri(fbo_texture->GetID(), GL_TEXTURE_WRAP_S, GL_CLAMP_TO_EDGE);
-	glTextureParameteri(fbo_texture->GetID(), GL_TEXTURE_WRAP_T, GL_CLAMP_TO_EDGE);
-}
+LightDrawer::LightDrawer() = default;
+LightDrawer::~LightDrawer() = default;
 
 void LightDrawer::draw(const ViewState& view, bool fog, const LightBuffer& light_buffer, const DrawColor& global_color, float light_intensity, float ambient_light_level) {
-	if (!shader) {
-		initRenderResources();
+	// Lazy-init resources on first draw
+	if (!shader_) {
+		shader_ = std::make_unique<LightShader>();
 	}
-
-	if (!fbo) {
-		InitFBO();
+	if (!fbo_) {
+		fbo_ = std::make_unique<LightFBO>();
 	}
 
 	int buffer_w = view.screensize_x;
@@ -84,16 +44,10 @@ void LightDrawer::draw(const ViewState& view, bool fog, const LightBuffer& light
 		return;
 	}
 
-	// 1. Resize FBO if needed (ensure it covers the visible screen area)
-	if (buffer_width < buffer_w || buffer_height < buffer_h) {
-		// Re-create texture if we need to grow
-		fbo_texture = std::make_unique<GLTextureResource>(GL_TEXTURE_2D);
-		ResizeFBO(std::max(buffer_width, buffer_w), std::max(buffer_height, buffer_h));
-		glNamedFramebufferTexture(fbo->GetID(), GL_COLOR_ATTACHMENT0, fbo_texture->GetID(), 0);
-	}
+	// 1. Ensure FBO is large enough for the visible screen area
+	fbo_->EnsureSize(buffer_w, buffer_h);
 
-	// 2. Prepare Lights
-	// Filter and convert lights to GPU format
+	// 2. Filter and convert lights to GPU format
 	gpu_lights_.clear();
 	gpu_lights_.reserve(light_buffer.lights.size());
 
@@ -104,12 +58,11 @@ void LightDrawer::draw(const ViewState& view, bool fog, const LightBuffer& light
 		float map_pos_x = static_cast<float>(lx_px - view.view_scroll_x);
 		float map_pos_y = static_cast<float>(ly_px - view.view_scroll_y);
 
-		// Convert to Screen Pixels
 		float screen_x = map_pos_x / view.zoom;
 		float screen_y = map_pos_y / view.zoom;
 		float screen_radius = (light.intensity * TILE_SIZE) / view.zoom;
 
-		// Check overlap with Screen Rect
+		// Frustum culling
 		if (screen_x + screen_radius < 0 || screen_x - screen_radius > buffer_w || screen_y + screen_radius < 0 || screen_y - screen_radius > buffer_h) {
 			continue;
 		}
@@ -117,39 +70,24 @@ void LightDrawer::draw(const ViewState& view, bool fog, const LightBuffer& light
 		wxColor c = colorFromEightBit(light.color);
 
 		gpu_lights_.push_back({ .position = { screen_x, screen_y }, .intensity = static_cast<float>(light.intensity), .padding = 0.0f,
-								// Pre-multiply intensity here if needed, or in shader
 								.color = { (c.Red() / 255.0f) * light_intensity, (c.Green() / 255.0f) * light_intensity, (c.Blue() / 255.0f) * light_intensity, 1.0f } });
 	}
 
-	if (gpu_lights_.empty()) {
-		// Just render ambient? We still need to clear the FBO/screen area or simpy fill it.
-		// If no lights, the overlay should just be ambient color.
-	} else {
-		// Upload Lights
-		size_t needed_size = gpu_lights_.size() * sizeof(GPULight);
-		if (needed_size > light_ssbo_capacity_) {
-			light_ssbo_capacity_ = std::max(needed_size, static_cast<size_t>(light_ssbo_capacity_ * 1.5));
-			if (light_ssbo_capacity_ < 1024) {
-				light_ssbo_capacity_ = 1024;
-			}
-			// Destroy and recreate buffer for Immutable Storage
-			light_ssbo = std::make_unique<GLBuffer>();
-			glNamedBufferStorage(light_ssbo->GetID(), light_ssbo_capacity_, nullptr, GL_DYNAMIC_STORAGE_BIT);
-		}
-		glNamedBufferSubData(light_ssbo->GetID(), 0, needed_size, gpu_lights_.data());
+	// 3. Upload light data to SSBO
+	if (!gpu_lights_.empty()) {
+		shader_->Upload(gpu_lights_);
 	}
 
-	// 3. Render to FBO
+	// 4. Render to FBO
 	{
-		ScopedGLFramebuffer fboScope(GL_FRAMEBUFFER, fbo->GetID());
+		ScopedGLFramebuffer fboScope(GL_FRAMEBUFFER, fbo_->GetFBOID());
 		ScopedGLViewport viewportScope(0, 0, buffer_w, buffer_h);
 
-		// Clear to Ambient Color
+		// Clear to ambient color
 		float ambient_r = (global_color.r / 255.0f) * ambient_light_level;
 		float ambient_g = (global_color.g / 255.0f) * ambient_light_level;
 		float ambient_b = (global_color.b / 255.0f) * ambient_light_level;
 
-		// If global_color is (0,0,0) (not set), use a default dark ambient
 		if (global_color.r == 0 && global_color.g == 0 && global_color.b == 0) {
 			ambient_r = 0.5f * ambient_light_level;
 			ambient_g = 0.5f * ambient_light_level;
@@ -157,209 +95,25 @@ void LightDrawer::draw(const ViewState& view, bool fog, const LightBuffer& light
 		}
 
 		glClearColor(ambient_r, ambient_g, ambient_b, 1.0f);
-		// Actually, for "Max" blending, we want to start with Ambient.
 		glClear(GL_COLOR_BUFFER_BIT);
 
 		if (!gpu_lights_.empty()) {
-			shader->Use();
-
-			// Setup Projection for FBO: Ortho 0..buffer_w, buffer_h..0 (Y-down)
-			// This matches screen coordinate system and avoids flips
 			glm::mat4 fbo_projection = glm::ortho(0.0f, static_cast<float>(buffer_w), static_cast<float>(buffer_h), 0.0f);
-			shader->SetMat4("uProjection", fbo_projection);
-			shader->SetFloat("uTileSize", static_cast<float>(TILE_SIZE) / view.zoom);
-
-			glBindBufferBase(GL_SHADER_STORAGE_BUFFER, 0, light_ssbo->GetID());
-			glBindVertexArray(vao->GetID());
-
-			// Enable MAX blending
-			{
-				ScopedGLCapability blendCap(GL_BLEND);
-				ScopedGLBlend blendState(GL_ONE, GL_ONE, GL_MAX); // Factors don't matter much for MAX, but usually 1,1 is safe
-
-				if (gpu_lights_.size() > static_cast<size_t>(std::numeric_limits<GLsizei>::max())) {
-					spdlog::error("Too many lights for glDrawArraysInstanced");
-					return;
-				}
-				glDrawArraysInstanced(GL_TRIANGLE_FAN, 0, 4, static_cast<GLsizei>(gpu_lights_.size()));
-			}
-
-			glBindVertexArray(0);
+			float tileSize = static_cast<float>(TILE_SIZE) / view.zoom;
+			shader_->DrawLightPass(fbo_projection, tileSize, static_cast<int>(gpu_lights_.size()));
 		}
 	}
 
-	// 4. Composite FBO to Screen
-	// Actually MapDrawer doesn't seem to set viewport every time, but `view.projectionMatrix` assumes 0..screensize.
-
-	// Use PrimitiveRenderer or just a simple quad?
-	// We can use a simple Blit shader or re-use `sprite_drawer->glBlitSquare` if it supports textures.
-	// Or just do a quick manual draw using fixed pipeline or a simple shader.
-	// Let's use `glBlitNamedFramebuffer`? No, we need blending (Multiply).
-	// So we draw a textured quad.
-
-	// We can reuse the `shader` but we need a "PASS THROUGH" mode or a separate shader.
-	// Easier to just use `glEnable(GL_TEXTURE_2D)` and fixed function if compatible? No, we are in Core Profile likely.
-	// Let's assume we need a simple texture shader.
-	// BUT wait, `LightDrawer::draw` previously drew a quad with the computed light.
-	// We can just use a simple "Texture Shader" here.
-
-	// WARNING: We don't have a generic texture shader easily accessible here?
-	// `floor_drawer` etc use `sprite_batch`.
-	// We can use `sprite_batch` to draw the FBO texture!
-	// `sprite_batch` usually takes Atlas Region.
-	// Does `SpriteBatch` support raw texture ID?
-	// `SpriteBatch` seems to assume Atlas.
-
-	// Let's use the local `shader` with a "Mode" switch? Or just a second tiny shader.
-	// Adding a mode switch to `shader` is easiest. "uMode": 0 = Light Render, 1 = Composite.
-
-	shader->Use();
-	shader->SetInt("uMode", 1); // Composite Mode
-
-	// Bind FBO texture
-	glBindTextureUnit(0, fbo_texture->GetID());
-	shader->SetInt("uTexture", 0);
-
-	// Quad Transform for Screen
-	float draw_dest_x = 0.0f;
-	float draw_dest_y = 0.0f;
+	// 5. Composite FBO to screen
 	float draw_dest_w = static_cast<float>(buffer_w) * view.zoom;
 	float draw_dest_h = static_cast<float>(buffer_h) * view.zoom;
 
-	glm::mat4 model = glm::translate(glm::mat4(1.0f), glm::vec3(draw_dest_x, draw_dest_y, 0.0f));
+	glm::mat4 model = glm::translate(glm::mat4(1.0f), glm::vec3(0.0f, 0.0f, 0.0f));
 	model = glm::scale(model, glm::vec3(draw_dest_w, draw_dest_h, 1.0f));
-	shader->SetMat4("uProjection", view.projectionMatrix * view.viewMatrix * model); // reusing uProjection as MVP
+	glm::mat4 mvp = view.projectionMatrix * view.viewMatrix * model;
 
-	float uv_w = static_cast<float>(buffer_w) / static_cast<float>(buffer_width);
-	float uv_h = static_cast<float>(buffer_h) / static_cast<float>(buffer_height);
+	float uv_w = static_cast<float>(buffer_w) / static_cast<float>(fbo_->GetWidth());
+	float uv_h = static_cast<float>(buffer_h) / static_cast<float>(fbo_->GetHeight());
 
-	shader->SetVec2("uUVMin", glm::vec2(0.0f, uv_h));
-	shader->SetVec2("uUVMax", glm::vec2(uv_w, 0.0f));
-
-	// Blending: Dst * Src
-	{
-		ScopedGLCapability blendCap(GL_BLEND);
-		ScopedGLBlend blendState(GL_DST_COLOR, GL_ZERO);
-
-		glBindVertexArray(vao->GetID());
-		glDrawArrays(GL_TRIANGLE_FAN, 0, 4);
-		glBindVertexArray(0);
-	}
-
-	shader->SetInt("uMode", 0); // Reset
-}
-
-void LightDrawer::initRenderResources() {
-	// Modes: 0 = Light Generation (Instanced), 1 = Composite (Simple Texture)
-	const char* vs = R"(
-		#version 450 core
-		layout (location = 0) in vec2 aPos; // 0..1 Quad
-		
-		uniform int uMode;
-		uniform mat4 uProjection;
-		uniform float uTileSize;
-		uniform vec2 uUVMin;
-		uniform vec2 uUVMax;
-
-		struct Light {
-			vec2 position; 
-			float intensity;
-			float padding;
-			vec4 color;
-		};
-		layout(std430, binding = 0) buffer LightBlock {
-			Light uLights[];
-		};
-
-		out vec2 TexCoord;
-		out vec4 FragColor; // For Mode 0
-
-		void main() {
-			if (uMode == 0) {
-				// LIGHT GENERATION
-				Light l = uLights[gl_InstanceID];
-				
-				// Calculate quad size/pos in FBO space
-				// Radius spans 0..1 in distance math
-				// Quad size should cover the light radius
-				// light.intensity is in 'tiles'. 
-				// The falloff is 1.0 at center, 0.0 at radius = intensity * TILE_SIZE.
-				float radiusPx = l.intensity * uTileSize;
-				float size = radiusPx * 2.0;
-				
-				// Center position
-				vec2 center = l.position;
-				
-				// Vertex Pos (0..1) -> Local Pos (-size/2 .. +size/2) -> World Pos
-				vec2 localPos = (aPos - 0.5) * size;
-				vec2 worldPos = center + localPos;
-				
-				gl_Position = uProjection * vec4(worldPos, 0.0, 1.0);
-				
-				// Pass data to fragment
-				TexCoord = aPos - 0.5; // -0.5 to 0.5
-				FragColor = l.color;
-			} else {
-				// COMPOSITE
-				gl_Position = uProjection * vec4(aPos, 0.0, 1.0);
-				TexCoord = mix(uUVMin, uUVMax, aPos); 
-			}
-		}
-	)";
-
-	const char* fs = R"(
-		#version 450 core
-		in vec2 TexCoord;
-		in vec4 FragColor; // From VS
-		
-		uniform int uMode;
-		uniform sampler2D uTexture;
-
-		out vec4 OutColor;
-
-		void main() {
-			if (uMode == 0) {
-				// Light Falloff
-				// TexCoord is -0.5 to 0.5
-				float dist = length(TexCoord) * 2.0; // 0.0 to 1.0 (at edge of quad)
-				if (dist > 1.0) discard;
-				
-				float falloff = 1.0 - dist;
-				// Smooth it a bit?
-				// falloff = offset - (distance * 0.2)? Legacy formula:
-				// float intensity = (-distance + light.intensity) * 0.2f;
-				// Here we just use linear falloff for simplicity or match formula
-				// Visual approximation is fine for now.
-				
-				OutColor = FragColor * falloff;
-			} else {
-				// Texture fetch
-				OutColor = texture(uTexture, TexCoord);
-			}
-		}
-	)";
-
-	shader = std::make_unique<ShaderProgram>();
-	shader->Load(vs, fs);
-
-	float vertices[] = {
-		0.0f, 0.0f, // BL
-		1.0f, 0.0f, // BR
-		1.0f, 1.0f, // TR
-		0.0f, 1.0f // TL
-	};
-
-	vao = std::make_unique<GLVertexArray>();
-	vbo = std::make_unique<GLBuffer>();
-	light_ssbo = std::make_unique<GLBuffer>();
-
-	glNamedBufferStorage(vbo->GetID(), sizeof(vertices), vertices, 0);
-
-	glVertexArrayVertexBuffer(vao->GetID(), 0, vbo->GetID(), 0, 2 * sizeof(float));
-
-	glEnableVertexArrayAttrib(vao->GetID(), 0);
-	glVertexArrayAttribFormat(vao->GetID(), 0, 2, GL_FLOAT, GL_FALSE, 0);
-	glVertexArrayAttribBinding(vao->GetID(), 0, 0);
-
-	glBindVertexArray(0);
+	shader_->DrawComposite(mvp, fbo_->GetTextureID(), glm::vec2(0.0f, uv_h), glm::vec2(uv_w, 0.0f));
 }
